#4 #6 #7 Refactored: commands code moved to the command files
